+++ internal/data/tokens.go
@@ -2,14 +2,14 @@
 package data
 
 import (
-    "context"
-    "crypto/rand"
-    "crypto/sha256"
-    "database/sql"
-    "encoding/base32"
-    "time"
+	"context"
+	"crypto/rand"
+	"crypto/sha256"
+	"database/sql"
+	"encoding/base32"
+	"time"
 
-    "github.com/Joseph-Koop/json-project/internal/validator"
+	"github.com/Joseph-Koop/json-project/internal/validator"
 )
 
 // Purpose of the token
@@ -20,21 +20,20 @@ const ScopeAuthentication = "authentication"
 // Add struct tags. Only the token and the expiry time will be encoded
 // and sent in the JSON repsonse
 type Token struct {
-    Plaintext string      `json:"token"`     
-    Hash      []byte      `json:"-"`
-    UserID    int64       `json:"-"`
-    Expiry    time.Time   `json:"expiry"`
-    Scope     string      `json:"-"`
+	Plaintext string    `json:"token"`
+	Hash      []byte    `json:"-"`
+	UserID    int       `json:"-"`
+	Expiry    time.Time `json:"expiry"`
+	Scope     string    `json:"-"`
 }
 
-
 // Generate a token for the user
-func generateToken(userID int64, ttl time.Duration, scope string) (*Token, error) {
-    token := &Token {
-        UserID: userID,
-        Expiry: time.Now().Add(ttl),
-        Scope: scope,
-    }
+func generateToken(userID int, ttl time.Duration, scope string) (*Token, error) {
+	token := &Token{
+		UserID: userID,
+		Expiry: time.Now().Add(ttl),
+		Scope:  scope,
+	}
 	// Generate the actual token. We create a byte slice and fill it
 	// with random values (rand.Read)
 	randomBytes := make([]byte, 16)
@@ -44,26 +43,27 @@ func generateToken(userID int64, ttl time.Duration, scope string) (*Token, error
 	}
 	// Encode the random bytes using base-32
 	token.Plaintext = base32.StdEncoding.WithPadding(base32.NoPadding).EncodeToString(randomBytes)
-	// Now we hash the encoding. 
+	// Now we hash the encoding.
 	hash := sha256.Sum256([]byte(token.Plaintext))
-	token.Hash = hash[:]                    // array to slice conversion
+	token.Hash = hash[:] // array to slice conversion
 
 	return token, nil
 }
 
 // Validate the token the client sends back to us to be 26 bytes long
 func ValidateTokenPlaintext(v *validator.Validator, tokenPlaintext string) {
-    v.Check(tokenPlaintext != "", "token", "Must be provided.")
-    v.Check(len(tokenPlaintext) == 26, "token", "Must be 26 bytes long.")
+	v.Check(tokenPlaintext != "", "token", "Must be provided.")
+	v.Check(len(tokenPlaintext) == 26, "token", "Must be 26 bytes long.")
 }
 
 // Our access to the database
 type TokenModel struct {
-    DB *sql.DB
+	DB *sql.DB
 }
-// The New() method creates and returns a new token. It calls Insert() as a 
+
+// The New() method creates and returns a new token. It calls Insert() as a
 // helper method
-func (t TokenModel) New(userID int64, ttl time.Duration, scope string) (*Token, error) {
+func (t TokenModel) New(userID int, ttl time.Duration, scope string) (*Token, error) {
 	token, err := generateToken(userID, ttl, scope)
 	if err != nil {
 		return nil, err
@@ -75,12 +75,12 @@ func (t TokenModel) New(userID int64, ttl time.Duration, scope string) (*Token, 
 
 // Do the actual insert in to the database table
 func (t TokenModel) Insert(token *Token) error {
-    query := `
+	query := `
 		INSERT INTO tokens (hash, user_id, expiry, scope) 
 		VALUES ($1, $2, $3, $4)
 	`
 	args := []any{token.Hash, token.UserID, token.Expiry, token.Scope}
-	
+
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
 
@@ -89,15 +89,14 @@ func (t TokenModel) Insert(token *Token) error {
 }
 
 // Delete a token based on the type and the user
-func (t TokenModel) DeleteAllForUser(scope string, userID int64) error {
+func (t TokenModel) DeleteAllForUser(scope string, userID int) error {
 	query := `
 		DELETE FROM tokens 
 		WHERE scope = $1 AND user_id = $2
 	`
-    ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-    defer cancel()
+	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
+	defer cancel()
 
-    _, err := t.DB.ExecContext(ctx, query, scope, userID)
-    return err
+	_, err := t.DB.ExecContext(ctx, query, scope, userID)
+	return err
 }
-
